Fixes backup

+++ scripts/ntp/ntp_client-conf.sh
@@ -16,7 +16,9 @@ ntpdate 1.be.pool.ntp.org
 systemctl start ntp
 
 # Create a backup of the configuration file.
-cp /etc/ntp.conf /etc/ntp.bk
+if [ ! -f /etc/exports.bk ]; then
+    cp /etc/ntp.conf /etc/ntp.bk
+fi
 
 # Create the configuration file.
 echo "###### NTP CLIENT CONFIGURATION ######" > /etc/ntp.conf
